don't highlight trailing whitespace in insert mode
This takes a few lines from my .vimrc, which I got from some wiki, etc.
etc. so that trailing whitespace isn't highlighted unless one of the
following conditions is met:

  - There's trailing whitespace after the cursor, and we leave insert
    mode without getting rid of it
  - There's trailing whitespace after the cursor, but we move the cursor
    away from the trailing whitespace

This means that trailing whitespace isn't highlighted while we're in the
middle of typing a sentence, which is must less distracting while still
making it easy to notice styling issues in your code.

refs #4