add an option to apply style conditionnaly
Currently the kernel coding style applies to every buffers which has the
c or cpp filetype. Some users might want a finer control and apply this
style only to their kernels.

This commit changes the way the style is applied. Users can now define a
"g:linuxsty_patterns" array in their vimrc and the style will be applied
only if the buffer's path matches one of the pattern.